+++ pygbm/histogram.py
@@ -1,13 +1,5 @@
 import numpy as np
 from numba import njit
-from numba import float32, uint8
-from collections import namedtuple
-
-SplitInfo = namedtuple('SplitInfo', [
-    'gain', 'feature_idx', 'bin_idx',
-    'gradient_left', 'hessian_left',
-    'gradient_right', 'hessian_right',
-])
 
 
 HISTOGRAM_DTYPE = np.dtype([
@@ -64,57 +56,3 @@ def _build_histogram_unrolled(n_bins, sample_indices, binned_feature,
         histogram[bin_idx]['count'] += 1
 
     return histogram
-
-
-@njit(fastmath=False)
-def _split_gain(gradient_left, hessian_left, gradient_right, hessian_right,
-                gradient_parent, hessian_parent, l2_regularization):
-    """Loss reduction
-
-    Compute the reduction in loss after taking a split compared to keeping
-    the node a leaf of the tree.
-
-    See Equation 7 of:
-    XGBoost: A Scalable Tree Boosting System, T. Chen, C. Guestrin, 2016
-    https://arxiv.org/abs/1603.02754
-    """
-    def negative_loss(gradient, hessian):
-        return (gradient ** 2) / (hessian + l2_regularization)
-
-    gain = negative_loss(gradient_left, hessian_left)
-    gain += negative_loss(gradient_right, hessian_right)
-    gain -= negative_loss(gradient_parent, hessian_parent)
-    return gain
-
-
-@njit(locals={'gradient_left': float32, 'hessian_left': float32,
-              'best_gain': float32, 'best_bin_idx': uint8},
-      fastmath=True)
-def find_split(histogram, feature_idx, gradient_parent, hessian_parent,
-               l2_regularization):
-    gradient_left, hessian_left = 0., 0.
-    best_gain = -1.
-    for bin_idx in range(histogram.shape[0]):
-        gradient_left += histogram[bin_idx]['sum_gradients']
-        hessian_left += histogram[bin_idx]['sum_hessians']
-        gradient_right = gradient_parent - gradient_left
-        hessian_right = hessian_parent - hessian_left
-        gain = _split_gain(gradient_left, hessian_left,
-                           gradient_right, hessian_right,
-                           gradient_parent, hessian_parent,
-                           l2_regularization)
-        if gain >= best_gain:
-            best_gain = gain
-            best_bin_idx = bin_idx
-            best_gradient_left = gradient_left
-            best_hessian_left = hessian_left
-            best_gradient_right = gradient_right
-            best_hessian_right = hessian_right
-
-    return SplitInfo(best_gain,
-                     feature_idx,
-                     best_bin_idx,
-                     best_gradient_left,
-                     best_hessian_left,
-                     best_gradient_right,
-                     best_hessian_right)
